% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doble.R
\name{doble}
\alias{doble}
\title{Calcula el doble de un número, es decir, lo multiplica por 2}
\usage{
doble(x)
}
\arguments{
\item{x}{número}
}
\value{
El doble de \code{x}
}
\description{
Calcula el doble de un número, es decir, lo multiplica por 2
}
\examples{
doble(7)
doble(8)
}

