% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma dos números}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{número}

\item{y}{número}
}
\value{
La suma de \code{x} y \code{y}
}
\description{
Suma dos números
}
\examples{
suma.dos.numeros(2, 7)
suma.dos.numeros(3, 8)
}

